requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-s3-storage==0.13.2
djangorestframework==3.11.0
psycopg2-binary==2.8.5
google-api-python-client==1.12.8
gunicorn==19.8.1
celery==5.0.4
python-dotenv==0.8.2
redis==3.2.0
django-constance==2.8.0